Replaced a number of double float functions with their single precision float counterparts I did this only in places where only float32_t seems to be involved. Should not change a thing. Reason for change: code size reduction (a few kBytes). Speed improvements (marginal at best).

+++ mchf-eclipse/drivers/cat/cat_driver.c
@@ -1243,7 +1243,7 @@ static void CatDriver_HandleCommands()
             else
             {
                 // PWR / SWR
-                resp[0] =(limit_4bits(round(swrm.fwd_pwr))<<4)+limit_4bits(round(swrm.vswr_dampened));
+                resp[0] =(limit_4bits(roundf(swrm.fwd_pwr))<<4)+limit_4bits(roundf(swrm.vswr_dampened));
                 // ALC / MOD
                 resp[1] = (limit_4bits(0)<<4)+limit_4bits(0);
                 bc = 2;
@@ -1251,7 +1251,7 @@ static void CatDriver_HandleCommands()
 
             break;
         case FT817_READ_RX_STATE: /* E7 */
-            resp[0] = (uint8_t)round(sm.s_count*0.5);   //S-Meter signal
+            resp[0] = (uint8_t)roundf(sm.s_count*0.5);   //S-Meter signal
             bc = 1;
             break;
         case FT817_PTT_STATE:
